Das U-Boot: Fork U-Boot v2024.10 to support OreSat am335x custom boards.
Edit arch/arm/mach-omap2/Kconfig to include board/psas/oresat board
files. Based on the am335x_evm section of this file.

Copy board/ti/ to board/psas/. This includes the common directory and
is the template for the oresat family of boards.

Edit board/psas/oresat/Kconfig to use the correct files and to source
board/psas/common/Kconfig during build.

Modify board.c board.h and mux.c in board/psas/oresat to boot oresat
specific boards from the board and version strings found in the
EEPROM. This commit also comments out the replaced code and preserves
it for future reference.

Add configs/am335x_oresat_defconfig. This is the base configuraton
file u-boot uses to build images. CONFIG_OF_LIST is commented out
since it is not used in the current configuration but might be used
in the future.

Copy arch/arm/dts/am335x-evm.dts to arch/arm/dts/am335x-oresat.dts and
custommize the file. This is the default .dts u-boot uses on startup.
Set CONFIG_DEFAULT_DEVICE_TREE="am335x-oresat" inam335x_oresat_defconfig.
This .dtb is later replaced from the linux file system in /boot/dtbs/.

Copy include/configs/am335x_evm.h to include/configs/am335x_oresat.h
and edit the script to determine which .dtb to load. This script is
embedded in a macro as a string and passed the to the u-boot environment.

Add support for oresat-bbb-dev.dtb on the beaglebone black. The .dtb
is included in /boot/dtbs during the image builder process. Selection
takes place in include/configs/am335x_oresat.h. This .dtb was added
for testing purposes and can be used instead of am335x_boneblack.dtb
on the beaglebone black by modifying the board and version in the EEPROM.